// src/app/admin-dashboard/models/user.model.ts
export class User {
  'balance': {
    rvn: string;
    ltct: string;
    eth: string;
    btc: string;
  };
  'demoBalance': {
    rvn: number;
    ltct: number;
    eth: number;
    btc: number;
  };
  'temporary': {
    code: number;
  };
  'userName': string;
  'email': string;
  'phone': number;
  'activePlans': number;
  'activeDemoPlans': number;
  'devices': number;
  'createdAt': Date;
}