// src/app/admin-dashboard/models/userAsic.model.ts
export class UserAsic {
  '_id': string;
  'hostFees': number;
  'asicName': string;
  'asicStatus': boolean;
  'expired': boolean;
  'totalMined': number;
  'startDate': Date;
  'userID': string;
  'asicID': string;
  'createdAt': Date;
  'updatedAt': Date;
  'address': string;
  'workerID': string;
  'pool': string;
}